+++ internal/runners/clean/run_lin_mac.go
@@ -9,24 +9,65 @@ import (
 	"github.com/ActiveState/cli/internal/appinfo"
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/installation"
+	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/logging"
+	"github.com/ActiveState/cli/internal/output"
 )
 
-func removeConfig(cfg configurable) error {
+func (u *Uninstall) runUninstall() error {
+	// we aggregate installation errors, such that we can display all installation problems in the end
+	// TODO: This behavior should be replaced with a proper rollback mechanism https://www.pivotaltracker.com/story/show/178134918
+	var aggErr error
+	err := removeCache(u.cfg.CachePath())
+	if err != nil {
+		aggErr = locale.WrapError(aggErr, "uninstall_remove_cache_err", "Failed to remove cache directory {{.V0}}.", u.cfg.CachePath())
+	}
+
+	err = removeInstall(u.cfg, u.installDir)
+	if err != nil {
+		aggErr = locale.WrapError(aggErr, "uninstall_remove_executables_err", "Failed to remove all State Tool files in installation directory {{.V0}}", u.installDir)
+	}
+
+	err = removeConfig(u.cfg.ConfigPath(), u.out)
+	if err != nil {
+		aggErr = locale.WrapError(aggErr, "uninstall_remove_config_err", "Failed to remove configuration directory {{.V0}}", u.cfg.ConfigPath())
+
+	}
+
+	err = undoPrepare()
+	if err != nil {
+		aggErr = locale.WrapError(aggErr, "uninstall_prepare_err", "Failed to undo some installation steps.")
+	}
+
+	if aggErr != nil {
+		return aggErr
+	}
+
+	u.out.Print(locale.T("clean_success_message"))
+	return nil
+}
+
+func removeConfig(configPath string, out output.Outputer) error {
 	file, err := os.Open(logging.FilePath())
 	if err != nil {
-		return err
+		return locale.WrapError(err, "err_clean_open_log", "Could not open logging file at: {{.V0}}", logging.FilePath())
 	}
 	err = file.Sync()
 	if err != nil {
-		return err
+		return locale.WrapError(err, "err_clean_sync", "Could not sync logging file")
 	}
 	err = file.Close()
 	if err != nil {
-		return err
+		return locale.WrapError(err, "err_clean_close", "Could not close logging file")
 	}
 
-	return os.RemoveAll(cfg.ConfigPath())
+	err = os.RemoveAll(configPath)
+	if err != nil {
+		return locale.WrapError(err, "err_clean_config_remove", "Could not remove config directory")
+	}
+
+	out.Print(locale.Tl("clean_config_succes", "Successfully removed State Tool config directory"))
+	return nil
 }
 
 func removeInstall(_ configurable, installDir string) error {